Added size_t cast to disambiguate array usage.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@16946 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/x11/textctrl.cpp
@@ -1071,7 +1071,7 @@ void wxTextCtrl::SearchForBrackets()
     char bracket = ' ';
     
     if (m_cursorX > 0)
-        bracket = current[m_cursorX-1];
+        bracket = current[(size_t) m_cursorX-1];
         
     if (bracket == ')' || bracket == ']' || bracket == '}')
     {
